P3: define exporter ctype byte domains
Exporter name and unit suffix sanitizers passed plain char bytes directly to isalnum(). High-bit bytes are source-reachable through UTF-8-preserving RRD strings and raw chart or dimension IDs, so signed-char builds currently violate the C ctype argument contract. A concrete crash, memory corruption, or visible misclassification remains libc- and input-dependent.

Convert the classified byte to uint8_t at the two ctype boundaries. This supplies values in the required unsigned-char domain without changing the copied byte, sanitizer grammar, bounds, termination, return values, special unit rewrites, allocation, locking, or control flow.

The exporting_name_copy path covers all twelve chart and dimension calls in the collected and stored Graphite, OpenTSDB Telnet, and OpenTSDB HTTP formatters. The prometheus_units_copy path covers Prometheus text exposition and Prometheus remote write. Graphite plaintext and HTTP, OpenTSDB Telnet and HTTP JSON, Prometheus text, and remote-write output remain byte-identical for every defined pre-change input.

The fix removes reachable undefined behavior with no API, ABI, configuration, capability, concurrency, lifetime, or meaningful performance impact. It adds only a constant-time integer conversion per classified byte.

Validation compiled both owning translation units with the generated CMake commands in default, forced-signed-char, and forced-unsigned-char modes. ASan/UBSan byte-domain tests passed for every non-NUL byte and exporter-specific rewrites, git diff checks passed, and the final six-reviewer gate approved the exact diff with a clean raw-process audit.

(cherry picked from commit 3082b86b095c5468f54b2772163f8baec9b31655)

+++ src/exporting/process_data.c
@@ -19,7 +19,7 @@ size_t exporting_name_copy(char *dst, const char *src, size_t max_len)
     for (n = 0; *src && n < max_len; dst++, src++, n++) {
         char c = *src;
 
-        if (c != '.' && !isalnum(c))
+        if (c != '.' && !isalnum((uint8_t)c))
             *dst = '_';
         else
             *dst = c;
